% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{fix.fct}
\alias{fix.fct}
\title{Fix NULL values in dataframe}
\usage{
fix.fct(x)
}
\arguments{
\item{x}{Am object, possibly NULL}
}
\value{
A single object
}
\description{
Fix NULL values in dataframe
}
\examples{

x <- NULL
x2 <- fix.fct(x)
}
