% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_utils.R
\name{gdfpd.download.file}
\alias{gdfpd.download.file}
\title{Downalods files from the internet}
\usage{
gdfpd.download.file(dl.link, dest.file, max.dl.tries)
}
\arguments{
\item{dl.link}{Link to file}

\item{dest.file}{= Destination, as local file}

\item{max.dl.tries}{Maximum number of attempts for dowloading files}
}
\value{
Nothing
}
\description{
Downalods files from the internet
}
\examples{

my.url <- paste0('http://www.rad.cvm.gov.br/enetconsulta/',
                  'frmDownloadDocumento.aspx?CodigoInstituicao=2',
                  '&NumeroSequencialDocumento=46133')

\dontrun{ # keep CHECK fast
dl.status <- gdfpd.download.file(my.url, 'tempfile.zip', 10)
}
}
