% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_fct_get_files_from_bovespa.R
\name{gdfpd.get.files.from.bovespa}
\alias{gdfpd.get.files.from.bovespa}
\title{Fetches ALL new files from Bovespa}
\usage{
gdfpd.get.files.from.bovespa(my.id)
}
\arguments{
\item{my.id}{Company's ID}
}
\value{
A dataframe with several information about files
}
\description{
Fetches ALL new files from Bovespa
}
\examples{

 \dontrun{
 df.files <- gdfpd.get.files.from.bovespa(9512)
 }
}
