% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_read_dfp_zip_file.R
\name{gdfpd.read.dfp.zip.file}
\alias{gdfpd.read.dfp.zip.file}
\title{Reads a single zip file downloaded from Bovespa}
\usage{
gdfpd.read.dfp.zip.file(my.zip.file, folder.to.unzip = tempdir(),
  id.type)
}
\arguments{
\item{my.zip.file}{Full path to zip file}

\item{folder.to.unzip}{Folder to unzip files (default = tempdir())}

\item{id.type}{The type of file structure ('after 2011' or 'before 2011')}
}
\value{
A list with several dataframes containing financial statements
}
\description{
Reads a single zip file downloaded from Bovespa
}
\examples{

my.f <- system.file('extdata/9512_PETR_2002-12-31.zip', package = 'GetDFPData')

#my.l <- gdfpd.read.dfp.zip.file(my.f, id.type = 'before 2011')
#print(my.l)
}
