% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_read_dfp_zip_file.R
\name{gdfpd.read.dfp.zip.file.type.2}
\alias{gdfpd.read.dfp.zip.file.type.2}
\title{Reads folder for zip file pre 2011 (internal)}
\usage{
gdfpd.read.dfp.zip.file.type.2(rnd.folder.name,
  folder.to.unzip = tempdir())
}
\arguments{
\item{rnd.folder.name}{Folder where unzipped files are available}

\item{folder.to.unzip}{Folder to unzip files (default = tempdir())}
}
\value{
A list with financial statements
}
\description{
Reads folder for zip file pre 2011 (internal)
}
\examples{
# no example (this functions not used directly)
}
