% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_utils.R
\name{gdfpd.read.fwf.file}
\alias{gdfpd.read.fwf.file}
\title{Reads FWF file from bovespa (internal)}
\usage{
gdfpd.read.fwf.file(my.f, flag.thousands)
}
\arguments{
\item{my.f}{File to be read}

\item{flag.thousands}{A flag for thousands values}
}
\value{
A dataframe with data
}
\description{
Reads FWF file from bovespa (internal)
}
\examples{

my.f <- system.file('extdata/DFPBPAE.001', package = 'GetDFPData')

df.assets <- gdfpd.read.fwf.file(my.f, flag.thousands = FALSE)
}
