% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_read_fca_zip_file.R
\name{gdfpd.read.zip.file.type.fca}
\alias{gdfpd.read.zip.file.type.fca}
\title{Reads folder for FCA zip file contents (internal)}
\usage{
gdfpd.read.zip.file.type.fca(rnd.folder.name,
  folder.to.unzip = tempdir())
}
\arguments{
\item{rnd.folder.name}{Folder where unzipped files are available}

\item{folder.to.unzip}{Folder to unzip files, default = tempdir()}
}
\value{
A list with FCA data
}
\description{
Reads folder for FCA zip file contents (internal)
}
\examples{
# no example (this functions is not used directly)
}
