% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_utils.R
\name{gdfpd.search.company}
\alias{gdfpd.search.company}
\title{Helps users search for a company name}
\usage{
gdfpd.search.company(char.to.search, cache.folder = "DFP Cache Folder")
}
\arguments{
\item{char.to.search}{Character for partial matching}

\item{cache.folder}{Folder to cache (save) all processed information. Default = file.path(getwd(),'DFP Cache Folder')}
}
\value{
Names of found companies
}
\description{
Helps users search for a company name
}
\examples{

\dontrun{ # dontrun: keep cran check fast
gdfpd.search.company('GERDAU')
}
}
