% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_fct_get_files_from_bovespa.R
\name{get_files}
\alias{get_files}
\title{Fetches files for different systens (INTERNAL)}
\usage{
get_files(my.id, type.fin.report)
}
\arguments{
\item{my.id}{Company id}

\item{type.fin.report}{type of financial report (dfp/itr/fre/fca)}
}
\value{
A dataframe
}
\description{
Fetches files for different systens (INTERNAL)
}
\examples{

\dontrun{
df.fre.files <- get_files(9512, type.fin.report = 'dfp')
}
}
