% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_utils.R
\name{my.merge.dfs.lists}
\alias{my.merge.dfs.lists}
\title{Merges (row wise) dataframes from different list, using names of dataframes as index}
\usage{
my.merge.dfs.lists(l.1, l.2)
}
\arguments{
\item{l.1}{First dataframe}

\item{l.2}{Second dataframe}
}
\value{
A list with binded dataframes (same names as l.1)
}
\description{
Merges (row wise) dataframes from different list, using names of dataframes as index
}
\examples{

l.1 <- list(x = data.frame(runif(10)) )
l.2 <- list(x = data.frame(runif(10)) )

l <- my.merge.dfs.lists(l.1, l.2)

}
