% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.board.composition}
\alias{xml.fct.board.composition}
\title{Reads XML data for board composition}
\usage{
xml.fct.board.composition(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for board composition
}
\examples{

# No example (INTERNAL)
}
