% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.capital}
\alias{xml.fct.capital}
\title{Reads XML data for capita}
\usage{
xml.fct.capital(x)
}
\arguments{
\item{x}{A list with capital summary data}
}
\value{
A dataframe
}
\description{
Reads XML data for capita
}
\examples{

# No example (INTERNAL)
}
