% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.capital.reduction}
\alias{xml.fct.capital.reduction}
\title{Reads XML data for capital reduction data}
\usage{
xml.fct.capital.reduction(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for capital reduction data
}
\examples{

# No example (INTERNAL)
}
