% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.compensation}
\alias{xml.fct.compensation}
\title{Reads XML data for compensation}
\usage{
xml.fct.compensation(x)
}
\arguments{
\item{x}{A list with compensation data}
}
\value{
A dataframe
}
\description{
Reads XML data for compensation
}
\examples{

# No example (INTERNAL)
}
