% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.compensation.summary}
\alias{xml.fct.compensation.summary}
\title{Reads XML data for compensation summary data}
\usage{
xml.fct.compensation.summary(x)
}
\arguments{
\item{x}{A list with compensation summary data}
}
\value{
A dataframe
}
\description{
Reads XML data for compensation summary data
}
\examples{

# No example (INTERNAL)
}
