% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.debt}
\alias{xml.fct.debt}
\title{Reads XML data for debt}
\usage{
xml.fct.debt(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for debt
}
\examples{

# No example (INTERNAL)
}
