% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.family.related.parts}
\alias{xml.fct.family.related.parts}
\title{Reads XML data for family related parts}
\usage{
xml.fct.family.related.parts(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for family related parts
}
\examples{

# No example (INTERNAL)
}
