% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.family.relations}
\alias{xml.fct.family.relations}
\title{Reads XML data for family relations}
\usage{
xml.fct.family.relations(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for family relations
}
\examples{

# No example (INTERNAL)
}
