% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.repurchases}
\alias{xml.fct.repurchases}
\title{Reads XML data for repurchases}
\usage{
xml.fct.repurchases(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for repurchases
}
\examples{

# No example (INTERNAL)
}
