% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.responsible}
\alias{xml.fct.responsible}
\title{Reads XML data for responsibles documents}
\usage{
xml.fct.responsible(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for responsibles documents
}
\examples{

# No example (INTERNAL)
}
