% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.splits.inplits}
\alias{xml.fct.splits.inplits}
\title{Reads XML data for splits/inplits data}
\usage{
xml.fct.splits.inplits(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for splits/inplits data
}
\examples{

# No example (INTERNAL)
}
