% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.stock.values}
\alias{xml.fct.stock.values}
\title{Reads XML data for stock value}
\usage{
xml.fct.stock.values(x)
}
\arguments{
\item{x}{A list with stock value data}
}
\value{
A dataframe
}
\description{
Reads XML data for stock value
}
\examples{

# No example (INTERNAL)
}
