% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.stocks.details}
\alias{xml.fct.stocks.details}
\title{Reads XML data for stock details}
\usage{
xml.fct.stocks.details(x)
}
\arguments{
\item{x}{A list with data}
}
\value{
A dataframe
}
\description{
Reads XML data for stock details
}
\examples{

# No example (INTERNAL)
}
