% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{xml.fct.transactions.related}
\alias{xml.fct.transactions.related}
\title{Reads XML data for transaction data}
\usage{
xml.fct.transactions.related(x)
}
\arguments{
\item{x}{A list with transaction data}
}
\value{
A dataframe
}
\description{
Reads XML data for transaction data
}
\examples{

# No example (INTERNAL)
}
