% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfpd_xml_fcts.R
\name{fix.fct}
\alias{fix.fct}
\title{Fix NULL values in dataframe}
\usage{
fix.fct(x, type.info = "character", format.date = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{Am object, possibly NULL}

\item{type.info}{Type of object}

\item{format.date}{Format of data, as string}
}
\value{
A single object
}
\description{
Fix NULL values in dataframe
}
\examples{

x <- NULL
x2 <- fix.fct(x)
}
